+++ packages/patternfly-4/react-core/src/experimental/components/index.ts
@@ -1,2 +1,3 @@
 export * from './Button';
 export * from './Badge';
+export * from './Drawer';